Wed, May 29, 2024, 3:04 PM +08:00

+++ interpreter/interpreter.py
@@ -19,8 +19,8 @@ def calculator(x, y, z):
 
 #run function
 #define value
-x1=int(fomula_set[0])
-z1=int(fomula_set[-1])
+x1=float(fomula_set[0])
+z1=float(fomula_set[-1])
 y1=fomula_set[1]
 
 result=calculator(x1, y1, z1)
